feat: add global filter and move UnauthorizedException to global
