+++ src/utils/exception-global.filter.ts
@@ -4,17 +4,36 @@ import {
   ExceptionFilter,
   HttpException,
   HttpStatus,
+  UnauthorizedException,
 } from '@nestjs/common';
-import { ValidDataException } from './validate-data-exception';
 
 @Catch()
 export class ExceptionGlobalFilter<T> implements ExceptionFilter {
   catch(exception: T, host: ArgumentsHost) {
+    const response = host.switchToHttp().getResponse();
+    if (exception instanceof UnauthorizedException) {
+      const errorMessage = {
+        success: false,
+        status: 401,
+        code: 'AUTH_11',
+        message: '유효하지않은 토큰',
+      };
+      response.status(401).json(errorMessage);
+      return;
+    }
     const status =
       exception instanceof HttpException
         ? exception.getStatus()
         : HttpStatus.INTERNAL_SERVER_ERROR;
 
-    const response = host.switchToHttp().getResponse();
+    console.log(exception);
+    const errorMessage = {
+      success: false,
+      status: 401,
+      code: '???',
+      message: '백앤드를 불러주세요',
+    };
+    response.status(status).json(errorMessage);
+    return;
   }
 }
